LOW MAX SPEED HIGH ACCELERATION INTERESTING RESULTS
trying to make it work where rockets can have a low top speed but high acceleration. does not work as intended. pushing for future reference if needed but likely to be removed.

max speed limiting currently only applies to the direction the rocket is facing. could adjust by making the max speed limit apply to the overall velocity MAGNITUDE, not just for the direction the rocket is facing