in game timer
added timers for time spent moving, time standing still, and total time. total time may sometimes be off by .01 seconds due to truncating. it was either that or have the timer sometimes display 0 seconds and 100 centiseconds instead of 1 second and 0 centiseconds. i can't be bothered to spend more time on this rn