added laser emitters
lasers as in actual hitscan beams, not star wars pew pew blasters. added functionality for said beams to go in the four cardinal directions. need to adjust LaserEmitter and RocketController (probably) so that rockets that are destroyed can still kill the player with its explosion. the way it is right now, the rockets are simply destroyed without exploding.

- VolTorian